requirements.txt
pandas
pyarrow
pylance
duckdb
sentence_transformers